Dynamic buffer size

+++ src/types/peek_args.rs
@@ -1,5 +1,5 @@
 #[derive(Clone)]
 pub struct PeekArgs {
     pub addr: u64,
-    pub size: u32,
+    pub size: usize,
 }
